requirements.txt
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.1.0
google-cloud-storage==2.18.0
google-cloud-firestore==2.11.1
httpx==0.28.1
beautifulsoup4==4.14.3
2captcha-python==1.5.1
requests==2.31.0
gspread==6.1.2
google-api-python-client==2.150.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
functions-framework==3.9.2
flask==3.0.3
google-cloud-error-reporting==1.11.1
google-cloud-vision==3.4.1
MarkupSafe==2.1.3
firebase-admin==6.5.0
